Management Pack/adapter_requirements.txt
vmware-aria-operations-integration-sdk-lib~=1.0.0
ping3==4.0.8